Pull resolving-related items closer together
